fix: (93) move members data file
